boiling oceans, config file relocated for docker
