+++ .babelrc
@@ -1,4 +1,6 @@
 {
     "presets": ["react", "env"],
-    "plugins": ["transform-object-rest-spread"]
+    "plugins": ["transform-object-rest-spread", ["module-resolver", {
+        "root": ["./config"]
+    }]]
 }
