+++ src/backend/index.js
@@ -5,7 +5,7 @@ import morgan from 'morgan';
 import bodyParser from 'body-parser';
 import api from './routes/api';
 import frontend from './routes/frontend';
-import configurator from './config/configurator';
+import configurator from './utils/configurator';
 let config = configurator.get();
 
 let app = express();
